internwork/chatbot/Sustainbility _cahtbot/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
PyPDF2>=3.0.0